requirements.txt
python-dotenv
openai
# Optional if falling back to LangChain + Ollama
langchain
langchain-community
langchain-ollama
